+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 209:
 
 * Faster http::string_to_field
 * async_echo supports move-only handlers
+* test::stream maintains a handler work guard
 
 --------------------------------------------------------------------------------